Update customer.php
https://github.com/condor2/Opencart_23xx/issues/34

+++ upload/system/library/cart/customer.php
@@ -142,7 +142,7 @@ class Customer {
 		return $this->address_id;
 	}
 
-	public function getBalance(): int {
+	public function getBalance(): float {
 		$query = $this->db->query("SELECT SUM(amount) AS total FROM " . DB_PREFIX . "customer_transaction WHERE customer_id = '" . (int)$this->customer_id . "'");
 
 		return (float)$query->row['total'];
